:recycle: refactor: Grass component 분리 및 적용
